Added introduction, questions now actually asked.
I added a real introduction to the game, which is in main.py
Also added a question-asking and answering mechanism. There's
logic to deal with answering questions
At this point, however, only the first rep asks a question.
My first thought is to make it a for-loop, but I'm not sure if
I'm going to do that before or after I make actual answers to the questions.