Added randomly-generated names.
Each of the five reps will now have randomly generated names, rather
than the ones I preset in the system. This introduces a new dependency
on the names package, so I made a new branch to throw in a setup.py
file and a bunch of other stuff before I merge it back in.

This commit isn't functional unless you have the `names` package
already installed. The next few commits will be me adding in stuff
from setuptools.